+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+- Remove the Open Source Star banner that prompted users to star the GitHub repository